Some fix

+++ go_client/desktop_exports/health_check.go
@@ -8,8 +8,8 @@ import (
 )
 
 //export StartHealthCheck
-func StartHealthCheck(period int, sendMetrics bool) {
-	healthcheck.StartHealthCheck(int32(period), sendMetrics)
+func StartHealthCheck(period int32, sendMetrics bool) {
+	healthcheck.StartHealthCheck(period, sendMetrics)
 }
 
 //export StopHealthCheck
