+++ go_client/kotlin_exports/health_check.go
@@ -6,8 +6,8 @@ import (
 	log "go_client/logger"
 )
 
-func StartHealthCheck(period int, sendMetrics bool) {
-	healthcheck.StartHealthCheck(int32(period), sendMetrics)
+func StartHealthCheck(period int32, sendMetrics bool) {
+	healthcheck.StartHealthCheck(period, sendMetrics)
 }
 
 func StopHealthCheck() {
